wip/account-nft-info update of project name
