--- src/main/java/com/example/demo/xrpl/account/api/model/AccountView.java
@@ -1,5 +1,0 @@
-package com.example.demo.xrpl.account.api.model;
-
-public class AccountView {
-
-}
