--- src/main/java/com/example/demo/xrpl/configuration/DataSourceConfiguration.java
@@ -1,21 +1,0 @@
-package com.example.demo.xrpl.configuration;
-
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import javax.sql.DataSource;
-
-@Configuration
-public class DataSourceConfiguration {
-
-    @Bean
-    public DataSource dataSource() {
-        return DataSourceBuilder.create()
-                .driverClassName("org.postgresql.Driver")
-                .url("jdbc:postgresql://localhost:5432/postgres")
-                .username("root")
-                .password("root")
-                .build();
-    }
-}
