--- src/main/java/com/example/demo/xrpl/configuration/properties/XrplConfigurationProperties.java
@@ -1,11 +1,0 @@
-package com.example.demo.xrpl.configuration.properties;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@Data
-@ConfigurationProperties(prefix = "xrpl.addresses")
-public class XrplConfigurationProperties {
-    private String clientUrl;
-    private String faucetUrl;
-}
