--- src/main/java/com/example/demo/xrpl/nft/domain/model/NFTID.java
@@ -1,4 +1,0 @@
-package com.example.demo.xrpl.nft.domain.model;
-
-public class NFTID {
-}
