--- src/main/java/com/example/demo/xrpl/nft/domain/model/NFTStatus.java
@@ -1,7 +1,0 @@
-package com.example.demo.xrpl.nft.domain.model;
-
-public enum NFTStatus {
-    SPARE,
-    RENTED,
-
-}
