--- src/main/java/com/example/demo/xrpl/nft/domain/model/OfferID.java
@@ -1,5 +1,0 @@
-package com.example.demo.xrpl.nft.domain.model;
-
-public class OfferID{
-
-}
