--- src/main/java/com/example/demo/xrpl/nft/domain/model/OfferType.java
@@ -1,7 +1,0 @@
-package com.example.demo.xrpl.nft.domain.model;
-
-public enum OfferType {
-    COLLATERAL,
-    COLLATERAL_FREE,
-    ANY,
-}
