--- src/main/java/com/example/demo/xrpl/nft/domain/model/RentalStatus.java
@@ -1,7 +1,0 @@
-package com.example.demo.xrpl.nft.domain.model;
-
-public enum RentalStatus {
-    IN_PROGRESS,
-    ONGOING,
-    FINISHED
-}
