--- src/main/java/com/example/demo/xrpl/nft/domain/model/RentalType.java
@@ -1,6 +1,0 @@
-package com.example.demo.xrpl.nft.domain.model;
-
-public enum RentalType {
-    COLLATERAL,
-    COLLATERAL_FREE,
-}
