--- src/main/java/com/example/demo/xrpl/nft/domain/service/NftDetailsService.java
@@ -1,19 +1,0 @@
-package com.example.demo.xrpl.nft.domain.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.xrpl.xrpl4j.client.JsonRpcClientErrorException;
-import org.xrpl.xrpl4j.client.XrplClient;
-import org.xrpl.xrpl4j.model.client.accounts.AccountNftsResult;
-import org.xrpl.xrpl4j.model.transactions.Address;
-
-@Service
-@RequiredArgsConstructor
-public class NftDetailsService {
-
-    private final XrplClient xrplClient;
-
-    public AccountNftsResult getAll(Address address) throws JsonRpcClientErrorException {
-        return xrplClient.accountNfts(address);
-    }
-}
