+++ src/main/java/com/rentalSystem/xrpl/account/api/AccountInfoRestController.java
@@ -1,0 +1,24 @@
+package com.rentalSystem.xrpl.account.api;
+
+import com.rentalSystem.xrpl.account.domain.service.AccountService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.xrpl.xrpl4j.client.JsonRpcClientErrorException;
+import org.xrpl.xrpl4j.model.client.accounts.AccountInfoResult;
+import org.xrpl.xrpl4j.model.transactions.Address;
+
+@RestController
+@RequestMapping("account")
+@RequiredArgsConstructor
+class AccountInfoRestController {
+
+    private final AccountService accountService;
+
+    @GetMapping
+    AccountInfoResult getAccountInfo(@RequestParam String address) throws JsonRpcClientErrorException {
+        return accountService.getAccountInfo(Address.of(address));
+    }
+}
