+++ src/main/java/com/rentalSystem/xrpl/account/api/model/AccountView.java
@@ -1,0 +1,5 @@
+package com.rentalSystem.xrpl.account.api.model;
+
+public class AccountView {
+
+}
