+++ src/main/java/com/rentalSystem/xrpl/configuration/DataSourceConfiguration.java
@@ -1,0 +1,21 @@
+package com.rentalSystem.xrpl.configuration;
+
+import org.springframework.boot.jdbc.DataSourceBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import javax.sql.DataSource;
+
+@Configuration
+public class DataSourceConfiguration {
+
+    @Bean
+    public DataSource dataSource() {
+        return DataSourceBuilder.create()
+                .driverClassName("org.postgresql.Driver")
+                .url("jdbc:postgresql://localhost:5432/postgres")
+                .username("root")
+                .password("root")
+                .build();
+    }
+}
