+++ src/main/java/com/rentalSystem/xrpl/configuration/properties/XrplConfigurationProperties.java
@@ -1,0 +1,11 @@
+package com.rentalSystem.xrpl.configuration.properties;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@Data
+@ConfigurationProperties(prefix = "xrpl.addresses")
+public class XrplConfigurationProperties {
+    private String clientUrl;
+    private String faucetUrl;
+}
