+++ src/main/java/com/rentalSystem/xrpl/nft/api/model/NFTokenDto.java
@@ -1,0 +1,12 @@
+package com.rentalSystem.xrpl.nft.api.model;
+
+import lombok.Data;
+
+import javax.annotation.Nullable;
+
+@Data
+public class NFTokenDto {
+    private String id;
+    @Nullable
+    private String uri;
+}
