+++ src/main/java/com/rentalSystem/xrpl/nft/domain/model/NFTID.java
@@ -1,0 +1,4 @@
+package com.rentalSystem.xrpl.nft.domain.model;
+
+public class NFTID {
+}
