+++ src/main/java/com/rentalSystem/xrpl/nft/domain/model/NFTStatus.java
@@ -1,0 +1,7 @@
+package com.rentalSystem.xrpl.nft.domain.model;
+
+public enum NFTStatus {
+    SPARE,
+    RENTED,
+
+}
