+++ src/main/java/com/rentalSystem/xrpl/nft/domain/model/NFTokenOffers.java
@@ -1,0 +1,11 @@
+package com.rentalSystem.xrpl.nft.domain.model;
+
+import lombok.Builder;
+import org.xrpl.xrpl4j.model.client.nft.BuyOffer;
+import org.xrpl.xrpl4j.model.client.nft.SellOffer;
+
+import java.util.List;
+
+@Builder
+public record NFTokenOffers(List<SellOffer> sellOffers, List<BuyOffer> buyOffers) {
+}
