+++ src/main/java/com/rentalSystem/xrpl/nft/domain/model/OfferID.java
@@ -1,0 +1,5 @@
+package com.rentalSystem.xrpl.nft.domain.model;
+
+public class OfferID{
+
+}
