+++ src/main/java/com/rentalSystem/xrpl/nft/domain/model/OfferType.java
@@ -1,0 +1,7 @@
+package com.rentalSystem.xrpl.nft.domain.model;
+
+public enum OfferType {
+    COLLATERAL,
+    COLLATERAL_FREE,
+    ANY,
+}
