+++ src/main/java/com/rentalSystem/xrpl/nft/domain/model/RentalStatus.java
@@ -1,0 +1,7 @@
+package com.rentalSystem.xrpl.nft.domain.model;
+
+public enum RentalStatus {
+    IN_PROGRESS,
+    ONGOING,
+    FINISHED
+}
