+++ src/main/java/com/rentalSystem/xrpl/nft/domain/model/RentalType.java
@@ -1,0 +1,6 @@
+package com.rentalSystem.xrpl.nft.domain.model;
+
+public enum RentalType {
+    COLLATERAL,
+    COLLATERAL_FREE,
+}
