+++ src/test/java/com/example/demo/XrplRentalSystemTests.java
@@ -1,0 +1,13 @@
+package com.example.demo;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class XrplRentalSystemTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
